SignUp functionality Completed
User enters data: Name, email, password, etc in signup page form.
State management: The data is stored at context of rtk  and the signup process is initiated.
OTP verification: An email with a one-time password (OTP) is sent for verification with nodemailer and that otp stored in db.
OTP validation: User enters the received OTP in verifymail page, which is then validated against the one stored in the database.
User creation: Upon successful validation, a new user is created using the provided data.